Ensure correct parser structure

This function checks that the passed object is a function and contains the arguments "x" and "..." - in that order. This is to help ensure users only register valid parsers.

Usage

validate_qryflow_parser(parser)

Arguments

parser

object to check

Value

Logical. Generates an error if the object does not pass all the criteria.

See also

Examples

custom_func <- function(x, ...){

  # Parsing Code Goes Here

}
validate_qryflow_parser(custom_func)